Move quote mismatch handling to Rules and add comments
